backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
pymupdf
Pillow
pytesseract
python-dotenv
pygments
raganything[text]
